feat: create ClaimsService
